requirements.txt
absl-py==2.1.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docstring_parser==0.16
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
etils==1.9.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.4
fsspec==2024.6.1
gast==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.36.0
imgaug==0.4.0
immutabledict==4.2.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kagglehub==0.2.8
keras==2.2.4
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multitasking==0.0.11
mysql-connector-python==9.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
patsy==0.5.6
peewee==3.17.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pmdarima==2.0.4
pooch==1.8.2
prettytable==3.10.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg==3.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==17.0.0
pyclipper==1.3.0.post6
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
scapy==2.5.0
Send2Trash==1.8.3
shapely==2.0.6
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.2
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
toml==0.10.2
torch==2.4.0
torchinfo==1.8.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.98
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.2.1
validators==0.34.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
